+++ src/main/java/com/nttdata/service/IClienteServicio.java
@@ -9,6 +9,6 @@ import com.nttdata.model.Cliente;
  * @author gustavoefrainparcosanchez
  *
  */
-public interface IClienteServicio extends ICRUD<Cliente, Integer> {
+public interface IClienteServicio extends ICRUD<Cliente, Long> {
 
 }
